Add C++ Solution to Patterns/Stack & Queue

+++ LeetCode/Patterns/Stack & Queue/155. Min Stack/Solution.cpp
@@ -17,7 +17,7 @@ public:
         else{
             if(value>min_ele) st.push(value);
             else{
-                st.push((2*value)%mod-min_ele);
+                st.push((long)(2*value)%mod-min_ele);
                 min_ele=value;
             }
         }
